+++ src/main/java/org/folio/bulkops/configs/CacheConfig.java
@@ -14,7 +14,7 @@ import java.util.concurrent.TimeUnit;
 public class CacheConfig {
   @Bean
   public Caffeine caffeineConfig() { //NOSONAR
-    return Caffeine.newBuilder().expireAfterWrite(60, TimeUnit.MINUTES);
+    return Caffeine.newBuilder().expireAfterWrite(5, TimeUnit.MINUTES);
   }
 
   @Bean
