+++ src/main/java/org/folio/bulkops/processor/ItemDataProcessor.java
@@ -8,29 +8,20 @@ import static org.folio.bulkops.domain.dto.UpdateActionType.REPLACE_WITH;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.PERMANENT_LOAN_TYPE;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.STATUS;
 
-import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Objects;
 
-import org.folio.bulkops.client.ConfigurationClient;
 import org.folio.bulkops.domain.bean.InventoryItemStatus;
 import org.folio.bulkops.domain.bean.Item;
 import org.folio.bulkops.domain.bean.ItemLocation;
 import org.folio.bulkops.domain.dto.Action;
 import org.folio.bulkops.domain.dto.UpdateOptionType;
 import org.folio.bulkops.exception.BulkOperationException;
-import org.folio.bulkops.exception.ConfigurationException;
 import org.folio.bulkops.exception.NotFoundException;
 import org.folio.bulkops.exception.RuleValidationException;
 import org.folio.bulkops.service.HoldingsReferenceService;
 import org.folio.bulkops.service.ItemReferenceService;
 import org.springframework.stereotype.Component;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -39,15 +30,8 @@ import lombok.extern.log4j.Log4j2;
 @Component
 @AllArgsConstructor
 public class ItemDataProcessor extends AbstractDataProcessor<Item> {
-
-  public static final String BULK_EDIT_CONFIGURATIONS_QUERY_TEMPLATE = "module==%s and configName==%s";
-  public static final String MODULE_NAME = "BULKEDIT";
-  public static final String STATUSES_CONFIG_NAME = "statuses";
-
   private final HoldingsReferenceService holdingsReferenceService;
   private final ItemReferenceService itemReferenceService;
-  private final ConfigurationClient configurationClient;
-  private final ObjectMapper objectMapper;
 
   @Override
   public Validator<UpdateOptionType, Action> validator(Item item) {
@@ -62,31 +46,13 @@ public class ItemDataProcessor extends AbstractDataProcessor<Item> {
         .getName()
         .getValue()
         .equals(action.getUpdated())
-          && !getAllowedStatuses(item.getStatus()
+          && !itemReferenceService.getAllowedStatuses(item.getStatus()
             .getName()
             .getValue()).contains(action.getUpdated())) {
         throw new RuleValidationException(
             format("New status value \"%s\" is not allowed", action.getUpdated()));
       }
     };
-  }
-
-  private List<String> getAllowedStatuses(String statusName) {
-    var configurations = configurationClient
-      .getConfigurations(format(BULK_EDIT_CONFIGURATIONS_QUERY_TEMPLATE, MODULE_NAME, STATUSES_CONFIG_NAME));
-    if (configurations.getConfigs()
-      .isEmpty()) {
-      throw new NotFoundException("Statuses configuration was not found");
-    }
-    try {
-      var statuses = objectMapper.readValue(configurations.getConfigs()
-        .get(0)
-        .getValue(), new TypeReference<HashMap<String, List<String>>>() {
-        });
-      return statuses.getOrDefault(statusName, Collections.emptyList());
-    } catch (JsonProcessingException e) {
-      throw new ConfigurationException(format("Error reading configuration, reason: %s", e.getMessage()));
-    }
   }
 
   @Override
