+++ src/test/java/org/folio/bulkops/processor/ItemDataProcessorTest.java
@@ -8,9 +8,9 @@ import static org.folio.bulkops.domain.dto.UpdateOptionType.PERMANENT_LOCATION;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.STATUS;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.TEMPORARY_LOAN_TYPE;
 import static org.folio.bulkops.domain.dto.UpdateOptionType.TEMPORARY_LOCATION;
-import static org.folio.bulkops.processor.ItemDataProcessor.BULK_EDIT_CONFIGURATIONS_QUERY_TEMPLATE;
-import static org.folio.bulkops.processor.ItemDataProcessor.MODULE_NAME;
-import static org.folio.bulkops.processor.ItemDataProcessor.STATUSES_CONFIG_NAME;
+import static org.folio.bulkops.service.ItemReferenceService.BULK_EDIT_CONFIGURATIONS_QUERY_TEMPLATE;
+import static org.folio.bulkops.service.ItemReferenceService.MODULE_NAME;
+import static org.folio.bulkops.service.ItemReferenceService.STATUSES_CONFIG_NAME;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
